+++ src/sn/CmnInterface.ts
@@ -135,6 +135,11 @@ export interface IVariable {
 	defValTrg(name: string, fnc: ISetVal): void;
 
 	doRecLog(): boolean;
+
+	get tagCh_doWait(): boolean;
+	get tagCh_doWait_Kidoku(): boolean;
+	get tagCh_msecWait(): number;
+	get tagCh_msecWait_Kidoku(): number;
 }
 export interface IValMp {[name: string]: string};
 export interface IValSave {[name: string]: string};